+++ src/commands/complex/arrow.ts
@@ -1,5 +1,4 @@
 import { LineOpts, line } from './line'
-import { beginPath, closePath, fill, setCtx, stroke } from '..'
 import { Point } from '../../types'
 import { RenderQueue } from '../../queue'
 import { arrowHead } from './arrow-head'
@@ -10,28 +9,28 @@ export type ArrowOpts = LineOpts & {
 
 export const arrow = (options: ArrowOpts) => {
 	return {
-		to: (queue: RenderQueue) => {
-			line(options).to(queue)
+		to: (q: RenderQueue) => {
+			line(options).to(q)
 
 			const prevPoint = options.points[options.points.length - 2] as Point
 			const lastPoint = options.points[options.points.length - 1] as Point
 
-			setCtx(options.options).to(queue)
-			beginPath().to(queue)
+			q.command.setCtx(options.options)
+			q.command.beginPath()
 
 			arrowHead({
 				tip: lastPoint,
 				d: { x: lastPoint.x - prevPoint.x, y: lastPoint.y - prevPoint.y },
 				length: options.headLength || 8,
-			}).to(queue)
+			}).to(q)
 
 			if (options.options.strokeStyle) {
-				stroke().to(queue)
+				q.command.stroke()
 			}
 			if (options.options.fillStyle) {
-				fill().to(queue)
+				q.command.fill()
 			}
-			closePath().to(queue)
+			q.command.closePath()
 		},
 	}
 }