+++ src/commands/complex/image.ts
@@ -1,13 +1,3 @@
-import {
-	arc,
-	beginPath,
-	clip,
-	closePath,
-	drawImage,
-	restore,
-	roundedRect,
-	save,
-} from '..'
 import { Box } from '../../types'
 import { RenderQueue } from '../../queue'
 
@@ -15,51 +5,51 @@ export type ImageOpts = {
 	img: string
 	box: Box
 	radius?: number
-	customShape?: (queue: RenderQueue, opts: ImageOpts) => void
+	customShape?: (q: RenderQueue, opts: ImageOpts) => void
 }
 
-export const roundedImage = (queue: RenderQueue, opts: ImageOpts) => {
+export const roundedImage = (q: RenderQueue, opts: ImageOpts) => {
 	const { box, radius = 0 } = opts
 	const { x, y } = box
 
-	beginPath().to(queue)
+	q.command.beginPath()
 
 	const halfWidth = box.w / 2
 
 	if (box.h === box.w && radius >= halfWidth) {
-		arc({ x: x + halfWidth, y: y + halfWidth, radius: halfWidth }).to(queue)
-		closePath().to(queue)
+		q.command.arc({ x: x + halfWidth, y: y + halfWidth, radius: halfWidth })
+		q.command.closePath()
 		return
 	}
 
-	roundedRect({ ...box, radius }).to(queue)
-	closePath().to(queue)
+	q.command.roundedRect({ ...box, radius })
+	q.command.closePath()
 }
 
 export const image = (options: ImageOpts) => {
 	return {
-		to: (queue: RenderQueue) => {
+		to: (q: RenderQueue) => {
 			let wasSaved = false
 
 			if (options.customShape) {
-				save().to(queue)
+				q.command.save()
 				wasSaved = true
-				options.customShape(queue, options)
-				clip().to(queue)
+				options.customShape(q, options)
+				q.command.clip()
 			} else if (options.radius) {
-				save().to(queue)
+				q.command.save()
 				wasSaved = true
-				roundedImage(queue, options)
-				clip().to(queue)
+				roundedImage(q, options)
+				q.command.clip()
 			}
 
-			drawImage({
+			q.command.drawImage({
 				img: options.img,
 				...options.box,
-			}).to(queue)
+			})
 
 			if (wasSaved) {
-				restore().to(queue)
+				q.command.restore()
 			}
 		},
 	}