+++ src/commands/primitives/clear-rect.ts
@@ -1,7 +1,3 @@
 import { Box } from '../../types'
-import { RenderCommand } from '../../queue'
-import { makeCommand } from '../make-command'
 
-export const clearRect = makeCommand<Box>(
-	(box: Box): RenderCommand => ({ command: 'clearRect', options: box })
-)
+export const clearRect = (box: Box): Box => box