+++ src/commands/primitives/clip.ts
@@ -1,4 +1,1 @@
-import { RenderCommand } from '../../queue'
-import { makeCommand } from '../make-command'
-
-export const clip = makeCommand<void>((): RenderCommand => ({ command: 'clip' }))
+export const clip = (): void => {}