+++ src/commands/primitives/close-path.ts
@@ -1,4 +1,1 @@
-import { RenderCommand } from '../../queue'
-import { makeCommand } from '../make-command'
-
-export const closePath = makeCommand<void>((): RenderCommand => ({ command: 'closePath' }))
+export const closePath = (): void => {}