+++ src/commands/primitives/draw-image.ts
@@ -1,11 +1,7 @@
 import { Box } from '../../types'
-import { RenderCommand } from '../../queue'
-import { makeCommand } from '../make-command'
 
 export type DrawImageOpts = Box & {
 	img: string
 }
 
-export const drawImage = makeCommand<DrawImageOpts>(
-	(options: DrawImageOpts): RenderCommand => ({ command: 'drawImage', options })
-)
+export const drawImage = (o: DrawImageOpts): DrawImageOpts => o