+++ src/commands/primitives/fill.ts
@@ -1,4 +1,1 @@
-import { RenderCommand } from '../../queue'
-import { makeCommand } from '../make-command'
-
-export const fill = makeCommand<void>((): RenderCommand => ({ command: 'fill' }))
+export const fill = (): void => {}