+++ src/commands/primitives/line-to.ts
@@ -1,7 +1,3 @@
 import { Point } from '../../types'
-import { RenderCommand } from '../../queue'
-import { makeCommand } from '../make-command'
 
-export const lineTo = makeCommand<Point>(
-	(p: Point): RenderCommand => ({ command: 'lineTo', options: p })
-)
+export const lineTo = (p: Point): Point => p