+++ src/commands/primitives/move-to.ts
@@ -1,7 +1,3 @@
 import { Point } from '../../types'
-import { RenderCommand } from '../../queue'
-import { makeCommand } from '../make-command'
 
-export const moveTo = makeCommand<Point>(
-	(p: Point): RenderCommand => ({ command: 'moveTo', options: p })
-)
+export const moveTo = (p: Point): Point => p