+++ src/commands/primitives/paste-bitmap.ts
@@ -1,13 +1,9 @@
 import { Box } from '../../types'
-import { RenderCommand } from '../../queue'
-import { makeCommand } from '../make-command'
 
 export type PasteBitmapOpts = Box & {
-	bitmap: string
+	bitmap: Uint8Array
 	length: number
 	channels: number
 }
 
-export const pasteBitmap = makeCommand<PasteBitmapOpts>(
-	(options: PasteBitmapOpts): RenderCommand => ({ command: 'pasteBitmap', options })
-)
+export const pasteBitmap = (o: PasteBitmapOpts): PasteBitmapOpts => o