+++ src/commands/primitives/rect.ts
@@ -1,7 +1,3 @@
 import { Box } from '../../types'
-import { RenderCommand } from '../../queue'
-import { makeCommand } from '../make-command'
 
-export const rect = makeCommand<Box>(
-	(box: Box): RenderCommand => ({ command: 'rect', options: box })
-)
+export const rect = (box: Box): Box => box