+++ src/commands/primitives/reset-transform.ts
@@ -1,4 +1,1 @@
-import { RenderCommand } from '../../queue'
-import { makeCommand } from '../make-command'
-
-export const resetTransform = makeCommand<void>((): RenderCommand => ({ command: 'resetTransform' }))
+export const resetTransform = (): void => {}