+++ src/commands/primitives/restore.ts
@@ -1,4 +1,1 @@
-import { RenderCommand } from '../../queue'
-import { makeCommand } from '../make-command'
-
-export const restore = makeCommand<void>((): RenderCommand => ({ command: 'restore' }))
+export const restore = (): void => {}