+++ src/commands/primitives/rotate.ts
@@ -1,6 +1,1 @@
-import { RenderCommand } from '../../queue'
-import { makeCommand } from '../make-command'
-
-export const rotate = makeCommand<number>(
-	(angle: number): RenderCommand => ({ command: 'rotate', options: angle })
-)
+export const rotate = (angle: number): number => angle