+++ src/commands/primitives/scale.ts
@@ -1,7 +1,3 @@
 import { Point } from '../../types'
-import { RenderCommand } from '../../queue'
-import { makeCommand } from '../make-command'
 
-export const scale = makeCommand<Point>(
-	(p: Point): RenderCommand => ({ command: 'scale', options: p })
-)
+export const scale = (p: Point): Point => p