+++ src/commands/primitives/stroke.ts
@@ -1,4 +1,1 @@
-import { RenderCommand } from '../../queue'
-import { makeCommand } from '../make-command'
-
-export const stroke = makeCommand<void>((): RenderCommand => ({ command: 'stroke' }))
+export const stroke = (): void => {}