+++ src/commands/primitives/translate.ts
@@ -1,7 +1,3 @@
 import { Point } from '../../types'
-import { RenderCommand } from '../../queue'
-import { makeCommand } from '../make-command'
 
-export const translate = makeCommand<Point>(
-	(p: Point): RenderCommand => ({ command: 'translate', options: p })
-)
+export const translate = (p: Point): Point => p